redirect if not lang
redirect if not lang
