+++ src/app/helpers/getAbsoluteUrl.js
@@ -1,0 +1,15 @@
+function absoluteUrl (req, setLocalhost) {
+    var protocol = 'https:'
+    var host = req ? req.headers.host : window.location.hostname
+    if (host.indexOf('localhost') > -1) {
+      if (setLocalhost) host = setLocalhost
+      protocol = 'http:'
+    }
+  
+    return {
+      protocol: protocol,
+      host: host
+    }
+  }
+
+  export default absoluteUrl
